working token storage

+++ back-end/src/main/java/tn/pfeconnect/pfeconnect/user/UserRepository.java
@@ -6,7 +6,9 @@ import tn.pfeconnect.pfeconnect.user.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByEmail(String email);
+
+
 }
